Tweak error propagation, restructure for MVC architecture
Altered error propagation, not the uncertanties are added in quadrature
wherever intensities (such as pixel or spot intensities) are summed.

Also resutuctured the code, encapsulating most of the processing
instructions (formerly in main.py) into their own function, seperating
them from the graphing functions. This is in preparation for main.py to
become a controller/dispatcher, user_interface.py to become the
beginnings of a view and pattern_processor.py to become the script that
controls the primary processing flow.